+++ agent/listener/server.py
@@ -22,7 +22,7 @@ import database
 import math
 
 
-__VERSION__ = '2.0.1'
+__VERSION__ = '2.0.2'
 __STARTED__ = datetime.datetime.now()
 __INTERNAL__ = False